changed the primary and accent colors
It's kinda surprising and revolting that mkdocs material doesn't ship
with "good" and accessible colors. Most of the primary colors offered
out of the box are simply not accessible enough with the light and dark
colorschemes and have miserable WCAG contrast ratios.

I've tried to improve the primary and accent color choices but
mkdocs-material really needs a "better" overall colorscheme.