updated the section on checking dependencies
TIL that we can also use `hash` and `type -a` to check for the presence
of binaries in $PATH. Both of them are shell built in keywords and
compatible with POSIX shells like dash and busybox ash.

`hash` is interesting because we don't need to redirect stdout since
nothing is printed if successful.